cmake: use luadoc for functions exported from C
Documenting half of the Lua API in the awesomerc manpage and the other
half with luadoc, doesn't make much sense. I modified the output of
gendoc.lua to something that looks like lua with luadoc markup. That
file can then be processed by luadoc and we have all Lua documentation in
one place.

And yes, we're now pulling custom doxygen tags out of C source code to
generate annotated lua source code we never use, which is then mangled
again by luadoc to spit out a set of html files.