awesome-menu: match string mode and autocomplete
I'm made some simple modifications of awesome-menu.c
to achieve the following:

1. Match the string entered by the user with the whole search string.
So far, it only matches from the start of the string. For menu entries
like 'gksu synaptic', the default behaviour is very inconvenient,
because obviously, I don't want to enter the 'gksu' part every time.
Another example: I have iceweasel, iceape and icedove installed.
Entering 'wea' is enough to find 'iceweasel' exclusively if the whole
string is searched. 'ice', on the other hand, won't do much good.

2. If there's just one single match, select that one automatically.
This saves hitting tab once in the cases when the choice is obvious
(because there is only one).

Signed-off-by: Julien Danjou <julien@danjou.info>