awesomerc: use gmatch instead of deprecated gfind
a user in #awesome made me aware of a little quirk in lib/awful.lua
which renders the history of the prompt unusable when using Gentoo and
installing Lua without the USE-flag "deprecated". Lua states that
string.gfind has been replaced with string.gmatch and aborts the
function prompt_history_save(id). I attached a patch that replaces
id:gfind with id:gmatch, which so far has not revealed any problems.

Signed-off-by: Julien Danjou <julien@danjou.info>