Made the fair layout really fair
The fair layout had the same issue as the slave columns in tiling layout
in that all strips were filled maximally with cells up to the last, and
the last strip may have significantly fewer cells than the other strips.
In my mind, that's not fair.  The new strategy makes sure the numbers of
cells in any two strips differ by no more than one.

Signed-off-by: Julien Danjou <julien@danjou.info>