client: fixed client_isfixed
here is a little patch which fix the unresizeable window bug.  This bug is due
to some windows, like firefox, having XCB_SIZE_HINT_P_MAX_SIZE and
XCB_SIZE_HINT_P_MIN_SIZE on and all related values to 0, which makes
client_isfixed believe that the windows are of fixed size.

Hope this is the right way to fix it. Anyway now you know where this bug comes
from.

Signed-off-by: Julien Danjou <julien@danjou.info>