Basic support for manpages in multiple languages
On 09/04/2010, Julien Danjou <julien@danjou.info> wrote:
>
> Well, I cannot find anything to say against this patch, so I've merged
> it.
>

Thanks for the kind comments and for your quick reply, I'm glad you
find this contribution useful.

I also noticed that in the previous patch there is a line that was not
supposed to be there, since its purpose was to help debugging the
cmake file. My apologies for that: I have removed it and attached the
corresponding patch (against the already patched version).

 Signed-off-by: Diego Moreda <diego.plan9@gmail.com>

Signed-off-by: Julien Danjou <julien@danjou.info>